`<Search/>` - factor out search
