+++ app/src/index.css
@@ -1,69 +1,3 @@
-
-.wow {
-  position: relative;
-}
-.search{
-  position: absolute;
-  top: 100%;
-  left: 0;
-  z-index: 10;
-  .suggested-theorems{
-    background: white;
-
-    .theorem{
-      padding-top: 5px;
-    }
-  }
-
-  background: white;
-  padding: 10px;
-
-  .hypothesis, .goal{
-    margin-bottom: 10px;
-    input{
-      color: inherit;
-      border: none;
-      background: transparent;
-      outline: none;
-      &::placeholder {
-        color: inherit;
-        opacity: 0.6;
-      }
-    }
-  }
-
-  button{
-    border: none;
-    padding: 3px 12px;
-    text-transform: uppercase;
-    font-size: 11px;
-    font-weight: bold;
-    color: #3c3b3bbd;
-    border-radius: 2px;
-    margin-top: 10px;
-    margin-right: 10px;
-    cursor: pointer;
-    background: #f0f0f0;
-    &:hover{
-      background: #e7e7e7;
-    }
-  }
-}
-
-.-is-searched{
-  border: 2px solid green !important;
-}
-
-
-
-
-
-
-
-
-
-
-
 /*______________________ BODY: prettiness ______________________*/
 body *{
   box-sizing: border-box;
